Move repository DB away from the cache directory
Yes, the DB is a cache. However, it will be filled again (with the same
amount of entries) when the app is started the next time, so there isn't
much sense in removing this file when the user clears the cache. When the
file is deleted while the app is running, all sorts of issues arise.

This file shouldn't be backed up though, which is quite simple on Lollipop
and later. Not sure if Former Android versions even have proper built-in
backup functionality. Also not sure if there's a way to tell TitaniumBackup
and similar apps to exclude some files, but I hope that they respect at
least the official no_backup directory.
(cherry picked from commit bdcfa9e)