Add a little compatibilities system for ZIPs
Now ZIPs can specify which requirements they have. There's only one for
now: SDK24+ requires that the installer uses /data/user_de/0 for the
BASE_DIR. If further requirements come up in the future, ZIPs can
simply specify them and the installer will show a warning that it needs
to be updated.

It's important to note that this is asking for certain features, not
for specific versions.

(cherry picked from commit ff695eb)